Improved error messages

+++ lsql-core/src/main/java/com/w11k/lsql/converter/predefined/AtomicIntegerConverter.java
@@ -22,7 +22,7 @@ public class AtomicIntegerConverter extends Converter {
 
     @Override
     protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
-        return new AtomicInteger((int) rs.getInt(index));
+        return new AtomicInteger(rs.getInt(index));
     }
 
 }
